Adds eligibility check into the strategy
