+++ Java/com/gildedrose/AgeingRules.java
@@ -2,4 +2,6 @@ package com.gildedrose;
 
 public interface AgeingRules {
     void process(Item currentItem);
+
+    boolean eligableFor(Item currentItem);
 }
